Simplify component tree
